[chore #162831438] refactor views to separate files
